+++ tests/SIL.LCModel.Tests/SIL.LCModel.Tests.csproj
@@ -147,6 +147,7 @@
     <Compile Include="DomainServices\AnalysisGuessServicesTests.cs" />
     <Compile Include="DomainServices\AnalysisOccurrenceTests.cs" />
     <Compile Include="DomainServices\DataMigration\DataMigration7000072Tests.cs" />
+    <Compile Include="DomainServices\InterlinearTestBase.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="DomainImpl\AtomicPropertyTests.cs" />
     <Compile Include="DomainServices\BaseStyleInfoTests.cs" />
